move world render in separate folder
